FPSciTests reorg and cleanup
